Cleaned up hooks
